pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Debsourya Datta <91617309+debsouryadatta@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0.post1"
sqlmodel = "^0.0.21"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
beautifulsoup4 = "^4.12.3"
langchain-cohere = "^0.2.2"
groq = "^0.10.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
